+++ packages/stream_chat/lib/src/db/chat_persistence_client.dart
@@ -17,6 +17,11 @@ abstract class ChatPersistenceClient {
   /// Whether the connection is established.
   bool get isConnected;
 
+  /// The current user id to which the client is connected.
+  ///
+  /// Returns `null` if the client is not connected.
+  String? get userId;
+
   /// Creates a new connection to the client
   Future<void> connect(String userId);