+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+🔄 Changed
+
+- Updated `dio` dependency to `^5.2.0`.
+
 ## 6.2.0
 
 🐞 Fixed