+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 6.3.0
 
 🐞 Fixed